+++ src/Database/Factories/TagFactory.php
@@ -12,7 +12,7 @@ class TagFactory extends Factory
     public function definition()
     {
         return [
-            'title' => $this->faker->word(),
+            'name' => $this->faker->word(),
             'slug' => $this->faker->slug()
         ];
     }