+++ src/Database/Migrations/2023_05_29_120000_create_tags_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
         Schema::connection(config('briofy-tag.database.connection'))
             ->create('tags', function (Blueprint $table) {
                 config('briofy-tag.database.uuid') ? $table->uuid('id')->primary() : $table->id();
-            $table->string('title');
-            $table->string('slug');
+            $table->string('name');
+            $table->string('slug')->unique();
             $table->timestamps();
             $table->softDeletes();
         });