+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## [Unreleased] - 2023-08-21
+
+### Added
+ - Add `unique` to `slug` in `tags` table migration.
+ - Add generate slug in model creating event.
+
+### Changes
+ - Change `title` to `name` in `tags`.
+
 ## Unreleased - 2023-08-21
 
 ### Added